Update timer to run immediately on start
When you clicked to shoot once the game was initialized a bullet would not be
fired. This was because the timer.canRun function when running for the first
time would not return true. Only on cycle 2 of calling the function was it
possible for it to return true. Now it return true on the first call resulting
it the player being able to shoot immediately on start