Update Objects to use 'self' instead of 'this'
'this' changes what it refer to depending how its called and its scope. When
called with 'new' in a Object + Function constructor it refers to the instance
of Object.

For inner function for Objects this will not refer to the object which is
confusing. Its either 'window' or 'undefined'. Assiging 'this' to a variable is
one solution for this. As a guide line, 'self' should be used instead of 'this'
in all objects so I don't have to think about wheather 'this' works or not.